Struct google_freebase1_sandbox::MethodSearchCall

pub struct MethodSearchCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Search Freebase open data.

This method supports media download. To enable it, adjust the builder like this: .param("alt", "media").

A builder for the search method. It is not used directly, but through a MethodMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.methods().search()
             .add_without("aliquyam")
             .add_with("ea")
             .add_type("no")
             .stemmed(true)
             .spell("justo")
             .scoring("et")
             .query("et")
             .prefixed(true)
             .output("ipsum")
             .mql_output("Lorem")
             .add_mid("et")
             .limit(-70)
             .add_lang("aliquyam")
             .indent(true)
             .help("Lorem")
             .format("eos")
             .add_filter("erat")
             .exact(false)
             .encode("dolor")
             .add_domain("eirmod")
             .cursor(-43)
             .callback("amet")
             .as_of_time("no")
             .doit();

Methods

impl<'a, C, A> MethodSearchCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn doit(self) -> Result<Response>

Perform the operation you have build so far.

fn add_without(self, new_value: &str) -> MethodSearchCall<'a, C, A>

A rule to not match against.

Append the given value to the without query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

fn add_with(self, new_value: &str) -> MethodSearchCall<'a, C, A>

A rule to match against.

Append the given value to the with query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

fn add_type(self, new_value: &str) -> MethodSearchCall<'a, C, A>

Restrict to topics with this Freebase type id.

Append the given value to the type query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

fn stemmed(self, new_value: bool) -> MethodSearchCall<'a, C, A>

Query on stemmed names and aliases. May not be used with prefixed.

Sets the stemmed query property to the given value.

fn spell(self, new_value: &str) -> MethodSearchCall<'a, C, A>

Request 'did you mean' suggestions

Sets the spell query property to the given value.

fn scoring(self, new_value: &str) -> MethodSearchCall<'a, C, A>

Relevance scoring algorithm to use.

Sets the scoring query property to the given value.

fn query(self, new_value: &str) -> MethodSearchCall<'a, C, A>

Query term to search for.

Sets the query query property to the given value.

fn prefixed(self, new_value: bool) -> MethodSearchCall<'a, C, A>

Prefix match against names and aliases.

Sets the prefixed query property to the given value.

fn output(self, new_value: &str) -> MethodSearchCall<'a, C, A>

An output expression to request data from matches.

Sets the output query property to the given value.

fn mql_output(self, new_value: &str) -> MethodSearchCall<'a, C, A>

The MQL query to run againist the results to extract more data.

Sets the mql_output query property to the given value.

fn add_mid(self, new_value: &str) -> MethodSearchCall<'a, C, A>

A mid to use instead of a query.

Append the given value to the mid query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

fn limit(self, new_value: i32) -> MethodSearchCall<'a, C, A>

Maximum number of results to return.

Sets the limit query property to the given value.

fn add_lang(self, new_value: &str) -> MethodSearchCall<'a, C, A>

The code of the language to run the query with. Default is 'en'.

Append the given value to the lang query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

fn indent(self, new_value: bool) -> MethodSearchCall<'a, C, A>

Whether to indent the json results or not.

Sets the indent query property to the given value.

fn help(self, new_value: &str) -> MethodSearchCall<'a, C, A>

The keyword to request help on.

Sets the help query property to the given value.

fn format(self, new_value: &str) -> MethodSearchCall<'a, C, A>

Structural format of the json response.

Sets the format query property to the given value.

fn add_filter(self, new_value: &str) -> MethodSearchCall<'a, C, A>

A filter to apply to the query.

Append the given value to the filter query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

fn exact(self, new_value: bool) -> MethodSearchCall<'a, C, A>

Query on exact name and keys only.

Sets the exact query property to the given value.

fn encode(self, new_value: &str) -> MethodSearchCall<'a, C, A>

The encoding of the response. You can use this parameter to enable html encoding.

Sets the encode query property to the given value.

fn add_domain(self, new_value: &str) -> MethodSearchCall<'a, C, A>

Restrict to topics with this Freebase domain id.

Append the given value to the domain query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

fn cursor(self, new_value: i32) -> MethodSearchCall<'a, C, A>

The cursor value to use for the next page of results.

Sets the cursor query property to the given value.

fn callback(self, new_value: &str) -> MethodSearchCall<'a, C, A>

JS method name for JSONP callbacks.

Sets the callback query property to the given value.

fn as_of_time(self, new_value: &str) -> MethodSearchCall<'a, C, A>

A mql as_of_time value to use with mql_output queries.

Sets the as_of_time query property to the given value.

fn delegate(self, new_value: &'a mut Delegate) -> MethodSearchCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> MethodSearchCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.